Fit multiple datasets with different functions, and at least one common fitting parameter.
Example of use with 2 gaussian and the same sigma:
bestFit -f data1.dat,data2.dat -t "a*exp(-((x-x01)/sigma)**2)","b*exp(-((x-x02)/sigma)**2)" -p a,b,x01,x02,sigma -i 1,1,1,0,1.

To be added to README.rst

Use Qt4 backend for Matplotlib if available, as it enables to
change all the figure elements, such as title, axis, colors etc.
At the moment there is a bug in it (matplotlib 1.0.1).